+++ packages/cli/src/commands/push.js
@@ -38,7 +38,7 @@ class PushCommand extends Command {
       : path.join(pwd, args.pattern);
 
     if (isFolder(filePattern)) {
-      filePattern = path.join(filePattern, '**/*.{js,jsx,ts,tsx,html}');
+      filePattern = path.join(filePattern, '**/*.{js,jsx,ts,tsx,html,vue}');
     }
 
     const appendTags = stringToArray(flags['append-tags']);
